[Feat] ( Client ) Set KeyCallBack To Player2D In Player2D::Start Function

+++ GameEngine/Include/Component/ColliderBox2D.cpp
@@ -178,13 +178,12 @@ void CColliderBox2D::Save(FILE* pFile)
 	CColliderComponent::Save(pFile);
 
 	fwrite(&m_Info, sizeof(Box2DInfo), 1, pFile);
-	/*
-	int MeshNameLength = m_Mesh->GetName().length();
+	
+	int MeshNameLength = (int)m_Mesh->GetName().length();
 
 	fwrite(&MeshNameLength, sizeof(int), 1, pFile);
 
 	fwrite(m_Mesh->GetName().c_str(), sizeof(char), MeshNameLength, pFile);
-	*/
 }
 
 void CColliderBox2D::Load(FILE* pFile)
@@ -192,8 +191,6 @@ void CColliderBox2D::Load(FILE* pFile)
 	CColliderComponent::Load(pFile);
 
 	fread(&m_Info, sizeof(Box2DInfo), 1, pFile);
-
-	/*
 
 	SetInheritRotZ(true); // m_Transform->m_InheritRotz를 true로 세팅하는 것
 
@@ -208,8 +205,6 @@ void CColliderBox2D::Load(FILE* pFile)
 	fread(MeshName, sizeof(char), MeshNameLength, pFile);
 
 	m_Mesh = m_Scene->GetResource()->FindMesh(MeshName);
-	*/
-
 }
 
 bool CColliderBox2D::Collision(CColliderComponent* Dest)
